waf: remove JAVA configuration & compilation tasks
Java was only used in one project that uses lib-common and not in
lib-common itself.
We should not check for Java if we only want to use lib-common, we
should do it in the project that uses it.
We also have since removed Java support from that project so it is
useless now.

rip-it: 9fc384076dd45e7074f259f1ff2dc9de6e05ea2c