lib-common: import zcore from the tools and split up into zparser and zpycore.
zparser.py and zbehave.py files of the old zcore module are only used
when running the tests from the build system. So move them in the
appropriate directory and use them when running tests from the build
system.

Remove a lot of code only used by the buildbots in zparser.py.
We should easily be able to move the removed code to the buildbot
zparser module in rd-infra.

The other files, zunittest.py and util.py are moved into a new module
zpycore to avoid conflicting with the old zcore module.

Instead of manipulating the sys.path, which causes lot of issues with
pylint, use some symlinks to use zpycore.

Change-Id: I779202159eddd20c734abcfbee586a5a68d13022

rip-it: 8169613c1f625aa958cf132c63af6caebf3ed1f2