iop/cfolder: fix exponent handling
It was possible to pass a very large exponent, and the loop used to
compute the results would take ages.

The problem is not only that we did not check overflow, because there are
other special cases (x = 0, 1 or -1).

This bug was spotted with AFL.

Change-Id: I1a3543f73696c4c0a30981227d0924b5ddfcf9e1

rip-it: 145c46c8f9cd3e9e09cc81e05cf63ff1decaa438 uprooted