zchk/src: fix memtoxll_ext test compile time.
This test alone took 8s on gcc-6.3.0, and 30s on clang-7.0.0 to compile
on default profile.

The solution is to put the big macro that was expended a lot of times
into a separate function.

Now, it takes 0.35s on gcc-6.3.0 and 0.40s on clang-7.0.0.

Change-Id: I80579053207acb19454367d2f027f7bcd7f3a089

rip-it: 5f01ad5bc1412f11f9baeb028c70a0493fb948cd