[BACKPORT]qps-hat: fix abusive use of ALWAYS_INLINE.
Backport of patch a64404a3489ae77e17caa5136e5d710614586990 from 19.3.

Reason: we are still fixing stuff in QPS hats in 2018.3 and the code
move created by the patch makes the merge of 2019.1 into 2019.3 painful
and error-prone. If we do the same code moves as in the 2019.3 patches,
we should make the things easier. This patch is not a cherry-pick: it
repeats the code moves done by the original patch.

Differences with the original patch:

- This patch turns additional functions that were still exposed into
  private (static) functions.

- Some small 'static inline' functions are moved to qps-hat.c as well
  because we did not need them to be static inline.

- The 'ALWAYS_INLINE' on 'lookup' and 'update_path' in qps-hat.in.c were
  not removed because of a suspicion of negative impact on performance
  (./zchk qps_hat ~13% slower with P=default).

==== Original commit message follows ====

In qps-hat.h, a lot of functions were using ALWAYS_INLINE even though
they can be long and do some complicated things.

The solution is to move the implementation of these functions to the
compile unit qps-hat.c.

Before this patch, clang-7.0.0 took forever to compile zchk-hat.c, now
it takes less than a second.

Change-Id: I558c63864424a14200f7b6545aedbb8792fe7ff5
Refs: #68530
Priv-Id: 7061c2b7e5e9d5b0a528e4ebd0477ee13beb3051