iop-yaml: add code for YAML parsing
Add the basis of a file that will handle IOP <-> YAML conversions.
For the moment, only the parsing from YAML to an internal AST is
written, and unit-tested. It still has a lot of issues, which will be
improved in smaller, more readable commits :)

Also, add a README about the syntax, which will become obsolete quickly,
and will have to be updated when the code has settled down.

Change-Id: I7697c28d8fd938729ea7face9b72405374ed7640

rip-it: 4622184af157f4ba8771fca0d22e163c44d12e9e