iop-yaml: improve end position for seq and obj
Every parsed data has in addition where in the input the data started
and ended. However, the end position for seq and obj was the start of
the next data, instead of the end of its last element, which makes more
sense. This is now fixed.

Change-Id: I4b1534ad23d9f39fa6af6cc70cc891197fa2578a

rip-it: be7f8d6a4cd1a6408d0fad8c0a3a34d413cc8db4