httpc: TLS client side.
Implement TLS in http client. This commit provides the possibility to
use encrypted communications with our http client but it does not provide
the verification of the server certificate. So the client can connect to
any server that implement TLS, regardless of their certificate.

Change-Id: I5136a53d159c858c2aeb96bccb625d0539829b8a

rip-it: 1c7687c17b74bad6c11ed9fb31938662392003b1