httpc: TLS client side.
Implement TLS in http client. This commit provides the possibility to
use encrypted communications with our http client but it does not provide
the verification of the server certificate. So the client can connect to
any server that implement TLS, regardless of their certificate.

Change-Id: I5136a53d159c858c2aeb96bccb625d0539829b8a

Conflicts:
	core-t.iop.h
	core.iop.c

rip-it: fab904b29342eb20cc531434b4403494902b6e71 uprooted