ic: remove mem_fifo_pool from ic_g.
Instead of using a mem_fifo_pool to allocate the ic_msg_t objects and
data, use the libc allocator.

Using a mem_fifo_pool was done to speed up the allocation and release of
ic_msg_t objects and data.
However, the RPCs and allocations are not guarantee to follow the FIFO
mechanism and it has been to fragment the memory in case of async RPCs.

Change-Id: I7eb558fafec9f71ea869be42fd29297a267f66e8

rip-it: 8cd726f4db6c33a9f7c88f4d7715c1d095729cfd