net-socket: do not call sock_reuseaddr when using UDP
Function bindx did always call sock_reuseaddr to set SO_REUSEADDR
option. This behavior is unwanted when using UDP as it allows multiple
calls to bind with the same <IP:PORT> (which would be probably a misuse)
instead of returning an error.

Change-Id: I47de6f52a31b8cd6ab1c2c5a11f55582bd2af1e9

rip-it: 37610e165be9994230ba0bac1f278be7e8996126