iop-yaml: factorize some packing functions with iop-json
Some functions were re-used from iop-json, but can be factorized as the
code is the same. This, however, means that the flags for the YAML
packing will have to be the same as the one for JSON packing.
For the moment, we re-use the IOP_JPACK_* flags in YAML, but we should
probably create a new one in the future, to:
* remove the flags that do not make sense in YAML
* not depend on JPACK flags for an "YPACK" API.

Change-Id: I879846e56a165ded982aafa7660f4318af77d2e5

rip-it: 9b3b1c719936a9c461a29cc85481b1c25185923c