iop-yaml: fix packing of strings containing specific yaml chars
In YAML, strings may be left unquoted unless the syntax of the document
becomes ambiguous. For this reason, strings must be quoted if:

* it can be confused with a key: "foo: "
* it can be confused with a comment: "# foo"
* it can be confused with anchors/derefs: "&anchor", "*deref"
* it can be confused with a quoted string because it itself starts
  with a quote: '"a'

The rules are simplified compared to real YAML, where "a:b" is ok but
"a: b" is not, the simple presence of the ':' character mandates
quoting.

The different cases are tested but there are still issues due to the
re-use of the formatting done in JSON. For '"' or \n etc characters, we
would rather print them as we write them, instead of printing them in
unicode format.

Change-Id: Iba3cdfa95b193c7bfac471eddad606733f430d90

rip-it: f5a328faecf90bfa9a016c2392a074bb334a12f6