iop-yaml: fix packing/unpacking of escaped characters
We want to be able to parse properly quoted strings, so that:
* escaped codes are handled (\r, \n, etc).
* the quote character is escaped.
* unicode characters are escaped.

This is exactly what is done in the json packer. However, the code was
riddled with macros using macros using macros using implicit variables,
which is horrible to factorize.
So I added even more macros (yeah!) with this time explicit parameters,
so that I can expose this function and provide all needed parameters in
the call.
There was still issues in the error handling, which has been moved in a
small wrapper on this public function.

In addition, the YAML packer code (which was copied from JSON) did not
handle these escaped characters nicely, only printing them in unicode
style. This has been fixed for YAML, but not for JSON, because I do not
want to change the behavior for JSON.

Change-Id: Ic2bb7b87553078dbcac3f3956233f485d546fd2f

rip-it: 60193b3425f26b1127a15ee269ee5065abe8e79a