iop-yaml: check IOP constraints on unpacking
We cannot check the constraints in yaml_data_to_iop_field, because they
might depend on the number of elements on an array (and thus we would
have to redo the checks once the array is built). Instead, do the
checks in the union & struct unpacker.

Change-Id: Ic6569c1ec578730dcb1ff1497019f9fdf94634ea

rip-it: cbf4f2a72b4f9c8e4addea468a14d9475638797c