iop-yaml: fix default packing flags for YAML
Instead of letting the user choose the packing flags, expose an API that
fixes those flags to a default value. This is because:
* We want to have default, sane flags.
* As YAML will be used for public interfaces, we want to ensure that
  certain flags are set, such as SKIP_PRIVATE.

An API to specify flags is still exposed for tests, but is documented as
not to be used.

Change-Id: I2491accb52649cbb41c062f20ccbeb034e884f97

rip-it: 267b2a122547bf5d70c44b29cf2c583ba555c19b