iop-yaml: clean up dead code from copy of iop-json
The iop-yaml code was boostrapped from iop-json code, and some functions
kept as is as they might have been useful, and the design was not clear
yet.
Now that the dust has settled, remove the dead code, so that the
iop-yaml code is clean, and the future commits more readable.

Change-Id: I4ceeeca1c06003a19ab7a063e868f8c122c1abbf

rip-it: 17a486252bd57334afc64d7a274e64306e058411