iop-rpc: use hook modified hdr when proxying
Commit eef3ec1eec introduced the possibility for the pre hook to modify
the header. However, this change was lost if the message was proxied.
To fix this, we want to make sure the message is repacked if the header
was modified. To avoid doing this every time, a parameter is added to
the hook to indicate whether the header was modified.

Change-Id: I182a44fa57ec56235177af54247b3e40506a301d

rip-it: 74f0e15ab8d19401c8cd358e94a364bb2bfd17c5