iop-yaml: fix bug in absent fields check on classes
When a class is unpacked, we have to go through each parent fields to
verify that all mandatory fields have been specified. This is done by
iterating on the total number of fields, and changing the current class
when its fields are all checked.

There was a typo that is fixed, that caused issue when unpacking classes
that have more fields in the parent that a child. Some comments have
been added to help understand the code a bit more.

Change-Id: Ifb81ac6fd0f2eeb015e82b69b6c2213073943bd8

rip-it: 613284114349e525261c903c25275f1b23c916df