iop-rpc: do pre_hook before proxy logic
Ensure the pre hook is always called before anything else.
This changes the behavior for the dynamic proxy case:

before:
* dynproxy call on hdr
* if error: throw PROXY_ERROR
* call pre hook on hdr (which may no longer be the hdr we will use)

after:
* call pre hook on hdr
* dynproxy call on hdr
* if error: throw PROXY_ERROR

This is clearly not backward compatible in the case of a PROXY_ERROR: we
now at least call the pre hook before. But this might actually be more
desirable, as the pre hook can at least process the fact that this RPC
has been called, even if there are errors afterwards (as is done in all
non dynproxy cases).

The code is still a bit ugly with all the seal/unseal and the
modifications of the ic object, but I don't see how to really simplify
it further.

Change-Id: I4a3751bd6e029e71eecd87bb930eb1d49cd483e0

rip-it: 4b0d3ba0516f6b4799be58cffcf7170d0954a691