iop-yaml: improve error message with location of error
Improve the error message by printing out the line where the error
begins (ie the line where the yaml data that caused the error began).
Below it, display carets to show the yaml data causing the issue.
If the yaml data is spread on several lines, we only show where the yaml
data starts.

Also, add the filename in the error message if the unpacking is done
with a file.

Change-Id: I25398a2aae0782732ef3c2dff0dc719bfc6a2635

rip-it: d01f4c6de0b184b9d365513e8ac81ae29c565dcc