iop: rename 'iop_value_set_absent()' as 'iop_field_set_absent()'
As the field is acting like 'this' in the context of this function, it
sounded more appropriate to name it this way.

In addition, there is already a 'iop_value_t' union that has no connection
with this function, contrarily to what the name could make think.

Change-Id: I24bba10523bd431617cce739b9d71a7ff2c5ce54

rip-it: cbcd4a7e1cde047e82167c6517475a219b2ac759