iop: fix alignment issue in 'iop_value_set_here()'
When allocating IOP structs/unions, the function used 'mp_new' with type
'char'. This type has an alignment of one but the structs we allocate
probably expect a bigger alignment.

This code didn't trigger any bug because our allocators aren't so accurate
about alignment and use a minimum alignment of 8, but it could lead to
tricky bugs if we begin to allocate IOPs on different memory pools in the
future.

Note: we use an alignment of 8, which is basically the biggest alignment
an IOP field can have, because we don't store the alignment in our
structure descriptions and calculating it would be unnecessarily
expensive.

This patch also adds a TODO to tell that we could use the raw version of
the indicator, as the fields are supposed to be properly initialized
afterwards, but for now we keep memsetting the structure to avoid any
unwanted side effects in the code. That change is left for another patch.

Change-Id: I9209736529fae5182c7ace835381be0d31767591

rip-it: 648d15757ebb9eebdcc087a7327e51f91c0de084