iop-yaml: fix unpacking with class with no fields
The handling of missing fields was buggy if a class had no fields: we
could use this class instead of its parent to check for absent fields,
leading to invalid reads.

This is fixed with a loop to ensure the "st->fields[i]" is valid.

Change-Id: I55d814eaece581fd18c0b1781aff84a58016bd4b

rip-it: 4776981dafcb57d106d2e72431757012b15bcada