core-macros: fix multi-evaluation of 'vec' for tab_for_each_{ptr, entry}
Since the introduction of helpers "FOR_INSTR*", it's way easier to create
for-like macros that used heterogeneous variable types. Use this capacity
to get rid of the long-lived multi-evaluation of the vector pointer in
'tab_for_each_ptr' and 'tab_for_each_entry'.

Note: it appears that in some contexts, this new version works a bit worse
with blocks than it did before but it's more a rewriter issue than it is a
problem with the macro.

Change-Id: I2e59b4d1123b78ace5eee8f24d3f0c15a81a60d1

rip-it: 51c9cb84d24289cb24d287ea123861975eb2fbd2