iop-yaml: improve obj unpacking complexity
The previous algorithm was something like this:
* iterate on the yaml object key -> value map.
** if a key matches, handle it and mark it as "seen" in a bitmap
* iterate on the IOP object fields
* if the field has not been set in "seen", verify it can be set absent.

This is not optimal for multiple reasons:
* we iterate on a map to then iterate on the iop fields to find a match,
  leading to a O(n^2) complexity, instead of using the map to find the
  matches
* we iterate a second time and thus need the use of a bitmap, leading to
  complex code.

This was done like this copying iop-json, but here we have access to a
map in the yaml object, so we can improve it:

* iterate on the IOP object fields
* find a match using the yaml object map
* handle presence/absence of data

The only missing check is then to make sure the yaml object map do not
contain extra keys. This situation is detected by counting the number
of keys that were mapped. Finding out which key is extra is then kept in
a non optimal but cold function (we re-iterate on the map to find a
missing key).

This changes the behavior on some invalid data: when previously we would
detect an unknown key immediately, we now detect missing keys first. A
test had to be rewritten for this reason.

Change-Id: I5a97533e10e42a5c2b072e9b560d18e9e2b8e8fc

rip-it: 3227de11acfed8c29d7554f1c3a9d07e101a7c46