iopc: fix bug in IOP error handling
In case of syntax errors in IOP dependancies, we sometimes saw unexpected
error messages like:

 "error: enum field name `X` is used twice"

But when looking the enum, there were no such duplicate.

This error was caused by the function 'check_path_exists' (the debate
about why this function, named this way, actually compiles IOP files
happen subsequently), which made no difference between not finding a file
and finding a file but getting an error while compiling it.

As a consequence, errors seen while compiling the broken IOP dependencies
were not terminal, so the compilation went on until something else
happened.

In the case of our NR test, it was:

 1. The enum was already parsed and added in our context so parsing it a
    second time made us detect false duplicates.

 2. The error 1. wasn't terminal either so we ended up telling that we
    didn't find the file 'nr_enum_error_dep.iop' (but we actually did).

Change-Id: I83dbb2b4c2d33f639b39da2c6d3bfcbc85cfbf94

rip-it: 5ee735a7849a307c14f1e43af409ce711d0786a2