yaml: slightly change column number computation
Instead of keeping a pointer to the last newline character, point to the
first character of the line. This requires a +1 when computing the
column number, but removes the hack used to work on the very first line
(pointing behind the first character of the stream, which is kind of
UB).

Change-Id: Id9798c62a3a50ccc8463322158e9609f43d38167

rip-it: 4e62766390c395e49c887ef79ace99e0218d3f86