yaml: force use of UINT for positive numbers
There was a corner case in the yaml parser that could lead to the INT
type being used with a positive number, when parsing "-0".

This is modified so that we enforce:
* UINT is used for integers >= 0
* INT is used for integers < 0

This allows simplifying the iop unpacker, as we know that INT cannot be
unpacked into unsigned types.

Change-Id: I4c04148e7e9708a118ff09da27267d6b8dda12ba

rip-it: 36fd7e27a329f4c3bd7bb29fc0719f2f8ef09124