iop-yaml: handle unpacking list of structs
A sequence of objects was not correctly unpacking. To fix it, the code
is a bit reworked to properly go into the seq case in
yaml_data_to_iop_field for all types of fields.

At the same time, the struct/union handling is moved inside the obj
case, and the special "null => struct" handling is moved inside the
"null scalar" handling function.

This is much better in terms of code logic.

Unfortunately, there is kind of a duplication between
yaml_data_to_iop_field and yaml_data_to_typed_struct, but it cannot
really be factorized as yaml_data_to_typed_struct is the entry-point of
unpacking.

Change-Id: I54322e46f4c536c48f76695b09c262718dc30d64

rip-it: 94e22638c8aaa2d3139f2dc950869cd3c4c2bb30