iop-yaml: test packing/unpacking of enum values in integer format
For packing, if an integer value for an enum field as no matching
string, then the integer value is packed.

For unpacking, if an integer value is found, it is unpacked only if
valid.

This is different from the behavior of the JSON unpacker, which will
always allow any integer value to be unpacked into an enum value.

iop.h is slightly reworked to add a version of iop_enum_exists that
works with the iop_enum_t object.

Change-Id: I04fa4ce1b46b70e8b0abe10a04f4d54168f299fb

rip-it: 05d11813dc2cf66104a7b86fe808332f7b629b90