str: Fix memtod function, should not accept empty strings.
The behavior of the function memtod is
inconsistent with the other similar functions like memtollp or
memtoip.

Set errno to EINVAL and return 0, in memtod if the given memory to parse
has a len equal to 0.

As a consequence the function "lstr_to_double" doesn't accept empty
string anymore, an error is now thrown in this case.

Change-Id: Ib443fa4354ffcd510b6f42cb7d1eb0a710a6b82b

rip-it: ff5c1e7917c0e0586180025ec20fa7be22362ee6