iop-openapi: add schema for primitive iop types
Add names for the iop primitives types, so that they can be easily
referred in more complex types.
Unfortunately, OpenAPI uses the "JSON data schema" which does not seem
to understand that unsigned number exists. So, there is no valid data
schema for uint64... It will be typed as "integer", with a format
"uint64", which is not standard, but widely used.

Change-Id: I2cf7913f857e136b9be0a95bebfa36a7d09fce79

rip-it: bdf0b8efb79f67197c9ca116c87d54ac807066c7