iop-openapi: handle fields with class type
When a field type is that of a class, it means any children of the class
is a valid value for this field.

OpenAPI is *extremely* ambiguous on how this is supposed to work. It
mentions using the discriminator field to avoid being redundant, but
does not provide a thorough example, making this completely unclear. I
expect better from a spec that is trying to be formally written.

The safest way to handle this is to list every possible children in a
single oneOf, which makes it very verbose, but normally safe.

Change-Id: Idc22f5657e597d7fafe9c1f3958464cea3a5a102

rip-it: ee19aaee0f1978fa3a815eea77f1530401fb62c3