iop-openapi: remove default iop:XXX primitive schemas
Instead of using schemas for the primitives, directly inline their
schema everywhere.

This is more verbose, but will allow expressing the IOP constraints
directly in the schema (otherwise, it was not possible with the $ref
version), and will probably be clearer for the user.

Change-Id: Ieb8ddfaf3818b955db0c5715e02664479b3be1e5

rip-it: d38107d4638c96da86aed7cf1263f9b4213e8e46