iop-openapi: handle integer constraints on fields
handle the min, max & nonEmpty constraints on iop fields. There are
issues when those constraints are used on double values, which to my
knowledge never happens, but it is a bug nonetheless.
Also, the nonEmpty on non unsigned value is not handled.


Change-Id: I3dff6b158c401aaaca4105cb9e06a560380c30ec

rip-it: 7ea7802ef50df7631e81e26a76ed4cae3edad645