yaml: fix packing of strings colliding with inline json
In yaml, it is normally possible to inline JSON:

a: [1, 2]
a: { b: 1, c: 2 }

We do not handle this syntax, but in order to ensure our packing is
properly parsed and handled by other tools, we should respect the spec.
That means that if a string starts with a [ or a {, it should be escaped
to prevent parsing it as an inline json string.

Change-Id: Id0710eb9e2898a2d606a9d04402c79f208b840fe

rip-it: 44e0fee267f3ba92b8b93c82b652090d32b29e2f