iop-openapi: handle IOP doc on interfaces and RPCs
Inspect the RPC attrs to retrieve the stored documentation comments for
RPCs.

As those comments can be set in the objects' help themselves, we are
still missing a lot of them. This will be fixed in the next commit.

Change-Id: Ib60cf22a3b7a39daec80b1a13017acf28e1f7df2

rip-it: 238e51e5ae893910f4067a61b6d713b53351b5ae