iop-openapi: move tests in dedicated file
As the tests uses IOP files that are only built and used for tests,
they must be separated from the iop-openapi.c file. This forces us to
use the public API instead of being able to unit-tests private
functions. The difference is not huge, but does add the openapi header
in every test file.

To avoid adding a lot of schemas in one case, we still expose a
"testing" API to clear registered schemas.

Change-Id: Ia66be4fe2c8c7595d33c7eebcad29b8963ec6533

rip-it: 7ca30acd5cac6b4ae2edae3a9a7552c81c1c0757