Move top-level test stuff to a "tests" directory.
The purpose of this commit and all the ones that will follow is to reduce
the number of files at the root directory of the repository, so that we
have a nice landing page on the github project. It is also the occasion
for structuring the code better.

Let's start by creating a directory dedicated to tests, and move the
top-levels test stuff (zchk and ztst-*) in it. Other tests will follow.

Change-Id: I72048830c0d3926a8c43638937d28ba9b9039b3b

rip-it: 24ef45ab2e1e8097bf4f22f4e769a8f72697f3ab