yaml: fix handling of empty seq/obj
We used the null value to encode an empty seq or obj. This is not valid
YAML, which requires the use of the JSON syntax in that case...
We do not handle inline json so we have to add specific code for those
two special strings.

Change-Id: I5c8e303c8136dd7779034ed1a032e7b05ac72385

rip-it: c1b56e980e8a137167e5e56d6eeb33f1930e9e03