ztst-httpd: fix invalid type for port variable
The port to use for the http server was declared as a bool, which
obviously is a tad too small to store a port number.
I have no idea why this error existed, the server could not work so the
binary was useless.

Change-Id: I59c9097e6cbc910c477764e99e34d2d1694d52ee

rip-it: 8f05d1d8b33bf68109604fb09179afcb0acee439