iop-yaml: allow use of tags for unions
The code is modified to allow the use of tags with unions, which
only works if the tag matches the name of the union struct.

The code is reworked a bit to make it work, we no longer go straight to
the union function, so some error messages have been updated to make
sure we only say "struct" for IOP_T_STRUCT, and "union" for IOP_T_UNION.

Additionally, the tag forbidden error message has been improved to
explicit the type on which the tag cannot be used. To have a nice error
message, a new parameter to yaml_data_get_type is required.

Change-Id: I257f5a861139ad25887be6f8e502bda4b46dd0d6

rip-it: a1d903dfa88614b37d3630b7a1dc21fd9f5886c3