iopc: Fix name conflict in generated .iop.ts files
The generated .iop.ts files contains the import of a Package type, used
to load the JSON metadata file. However, this name can conflict with a
user-declared type named Package, which is exactly what happens in one
of our IOP file.

This conflict was not detected by Typescript, but this bug has been
fixed in 3.7.

To be compatible with Typescript 3.7, we must avoid the name conflict.
The type is renamed to a name which cannot be user-generated (IOP types
cannot start with an underscore).

Change-Id: I2a428ed256b16166e52c0e56532a220779eb7dd6

rip-it: ae5480fe537949332062df28ceaf3348ad52af9b