iop-openapi: do not add tags of filtered out interfaces
The OpenAPI generation is based on an IOP module, which can contain
multiple interfaces. RPCs can be whitelisted individually, in which case
some interfaces in the module may have no whitelisted RPCs. In that
case, we do not want to add a tag for interface, as:
* it won't be used and will just clutter the output
* it might expose an interface we do not want to expose

Change-Id: Ib621b926b432f6bbc0c9562ddd9a017b4e76e6fc

rip-it: 478296049fe7cdfce7fe812614c390b416298fe7