iop-openapi: remove min/max bounds for some integer types
Keep the min/max for 8 and 16 bits integers, but remove them for 32 and
64 bits signed integers. For 32 and 64 bits unsigned integers, only
keep the 0 minimum value.

This will make the output clearer, and remove constraints with
unreadable values that clutter the result. The format still indicates
which type of value is expected.

Change-Id: Ia61f06376254076b3bb18c82c0e241b7409ae70b

rip-it: 90085f3a6720afd0b7269e1946fc1296e7d27719