yaml: add some basis for presentation data
When using YAML, a common use-case will be to parse it, possibly modify
it, and repack it. In this process, we want to be able to keep
"presentation data" from the first parsing, so that the repacking can
include this information.

Presentation data includes everything that related to how the YAML data
is *shown*, but that has no influence on the final parsed data. This
will include:
* Comments
* Newlines
* includes
* anchors & aliases
* templating parameters
* override parameters
* and maybe more

As the data may be slightly modified before being repacked (for example,
for migration & backward compatibility purposes), we need this
presentation data to be somewhat resilient to changes in the data.

To handle this, we want to keep semantic meaning of the presentation
data. This is done by associating it to paths in the data. To bootstrap
this, we start be associating inline comments to path, but the code will
be heavily changed & improved in future commits.

In addition, this means that the yaml parser/repacker will act as a
somewhat defacto standard formatter. A binary to do just that will be
added soon.

Change-Id: Iba1e52b3f8ba0829d73cf131a540bbea34f4798f

rip-it: 0509ed177ab6c6bbfac3aabae01b70d31de0ed1c