yaml: parse flow syntax for objects
Same logic as for flow sequences. The parsing is a bit reworked so
that it returns a key, data pair, which allows building the flow
objects progressively.

Unit tests for all error and passing cases are added, which caught
some mistakes in the previous commit.

Change-Id: Ifa8b363af0f6f4eabad3dedce53905f7d9a45141

rip-it: 9bc13c2719c6f4269a81d9f7f650a2e68afea1ec