README: Document compilation using clang first.
This commit modifies the README so that it mainly documents compilation
with clang, which is straight-forward compared to the compilation with
gcc.

Compilation with gcc is documented in a dedicated sub-section.

Change-Id: If67c4b626b13c931ce0f6cda2d54f4fb075f9989
(cherry picked from commit 27531fdadc91120155d206b556564f5992734d86)

rip-it: 7230ccba0148ff69b30aa47c6f61513d5ea4fd59