yaml: handle prefixed comments in presentation data
The presentation handling is reworked, so that we can save both
prefix comments and inline comments:

 # Prefix
 a: 2 # inline

To make this work, we now keep track of presentation data for the *next*
node, instead of only the current one. When the path is modified, if we
have data for the next node, we save it as presentation data for the new
path.

The code is reworked so that it is easy during parsing to push a new
"path", which will handle this next/current node stuff on its own.
To make this a bit easier, we no longer have a special case for the path
of a key in the root object: we now also prefix a key with a dot.

Change-Id: I91fa5baee49162e7b462ae8f3b4c3b8f4717e08d

rip-it: a8f260b8ae9b713bc8f13e56b45e7f6702eb0c77