yaml: add presentation data to yaml & iop-yaml public APIs
Add presentation data as an out param for unpacking APIs in yaml &
iop-yaml, and add presentation data as an in param for packing APIs in
yaml & iop-yaml.

This leaves the possibility for the user to retrieve the presentation
data, and use it when repacking.

Change-Id: Ia88750c23397084074fb01e86422427b50c1ce48

rip-it: 7aec19bf22e082b901cad36519a3b3d5e8b0a286