yaml: save flow mode in presentation data, use when packing
Flow mode is saved when parsed, and re-used to pack the data. Everything
is packed in the same line.

There is an issue here, in that tags are not allowed in flow mode. If
when packing in flow mode, we end up with a data holding a tag, we
have no other possibility than to lose information silently, which is
not ok.

A good fix would be to fallback to normal packing when such a situation
arises, to fix in a future commit.

Change-Id: I58692123ee8f4008aec5b733870992c8524910f7

rip-it: 19ee8fb905f0a6003031d933839252537bb8e8c7