tools: add yamlfmt binary
Add a binary that loads a yaml input (either from stdin or from a file),
and repacks it. This servers two purposes:
* it can lint the input and return an error if the file is illformed
* it can be used as an auto-formatter

An ALE plugin for VIM will be added to use this binary to have direct
error reporting as well as auto-formatting for our YAML files.

Change-Id: I7934e0a0330c07d7df7fdd715e5c20f31ecd53da

rip-it: f233c99d89ac8893232e8827dcf1732c3cef9a29