yaml: improve packing by adding cursor state
The packing was done on a single state named (badly) do_indent, which
distinguished a single case: are we packing a data following an object
key or not.
This was enough to properly handle packing, as we only had to consider:
* scalars (always add in same line)
* sequences (always add in a new line)
* objects (add in same line after a sequence dash, otherwise in a
  new line.

However, with the addition of presentation data, this is no longer
enough. We can have inline comments, empty lines, etc, and we want to
pack every properly so that we don't leave extraneous spaces or missing
newlines.

For example, we used to always pack a sequence dash with a space: "- ",
which is no longer ok if the following data has a prefix comment:

"-<space>\n"
"  # prefix comment\n"
"  a: 2\n"

To handle all those cases, the "do_indent" is replaced by an enum
describing the current packing cursor state. This allows any -> any
state transitions, solving all the corner cases.

Change-Id: I13bfec38c999c22a508cbcf7eb07b670fd052ab7

rip-it: 2499591aacfa91b544baf16fea132bdef1522db4