yaml: track inclusion chain, use it to pretty print errors
In every yaml_parse_t object, an "included details" is now set if the
file was included from another one. This struct contains the pointer to
the parent parsing context, as well as the "!include XXX" yaml data.

In addition, yaml_span_t objects now keep a pointer to the relevant
yaml_parse_t object. This allows us to properly pretty print errors for
any yaml_span_t (as the final AST will have the included AST inlined,
so different yaml_span_t objects may depend on different yaml_parse_t
objects).

This removes the need to pass the yaml_parse_t env to the pretty print
function. However, the tag_span is now a pointer, so that we can always
guarantee that span->env is non NULL (and thus only need to set the one
in tag_span if there is a tag for the yaml data).

The pretty printed error looks like this:

<topfile>:line:col: error in included file
   key: !include foo/bar.yml
        ^^^^^^^^^^^^^^^^
foo/bar.yml:line:col: error in included file
       - !include baz.yml
         ^^^^^^^^^^^^^^^^
...
bad.yml:line:col: <error>
<error_detail>

This could be reworked depending on what is clearer, both for users and
for tools (yamlfmt).

Change-Id: If56916a131c707e92f59af8cb4628b8457c49b89

rip-it: e219782ff9c603813088a0aff3c441aed8a2564a