qps: fix freelist corruption on realloc
The consistency of the QPS blocks could get corrupted on some realloc
situations. A block may end up marked as having its previous block free,
while it isn't. Use of the freelist might end up using the wrong blocks,
and then... you get 0x1010101 in all your handles. Makes sense!

To describe the bug, a quick summary about QPS, or rather about QPS's
implementation of the TLSF allocator. Allocated blocks have an
associated header, which can indicate the status of the block. Notably,
two bits are used to indicate whether the block is used or free, and
whether the previous block is used or free.

Here is a buggy situation:
* First block is used, with size N
* Second block is free, with size 1
* Third block is used, with size M

In the qps->hdrs, this ends up as:

+---------------+---------------+--------------------------+
| size: N, USED | size: 1, FREE | size: N, PREV_FREE, USED |
+---------------+---------------+--------------------------+

The first block is reallocated to size N+1. As the next block is free
and (block_size + next_block_size) = (N + 1) >= asked_size, The second
block is removed from the freelist, and the first block size adjusted:

+-------------------------------+--------------------------+
|       size: N + 1, USED       | size: N, PREV_FREE, USED |
+-------------------------------+--------------------------+

The bug is previous here: we should update the PREV_FREE flag of the new
next block, but we didn't. This leads to a corruption of the invariants
of the allocator (clearly detected by the check_invariants routine).
What this ends up causing probably depends a lot on the access pattern
of the QPS. This is too big of a bug to not have been triggered
regularly, but as the direct access to the freelist is still fine,
unless we end up reallocating or using blocks around this corrupted
block, we probably are safe.

As for the 0x1010101 value, well, here's the backtrace:
 #0  qhat_flatten_leaf8 at qps-hat.in.c:110
 #1  qhat_set_path8 at qps-hat.in.c:268
 #2  qhat_set_path at qps-hat.h:286
We end up writing the byte value 1 repeatedly in a uint8_t array, at the
wrong address, with a wrong len.
My guess is that the bug is caused by repeated +1 reallocs, followed by
repeated freeing. We end up corrupting the headers, then using a wrong
header when reallocating, using an invalid ptr & len when setting a
simple value "1".  This ends up corrupting a handle's list with
repeated 0x1010101 value, which can get copied around, but leads
to segfaults when dereferenced.

Change-Id: I4650d9666c5cab3af8ab824b98a300b9095ead8a

rip-it: 39cffc8bcf9314bb731798a377ae97b2d77664f7 uprooted