yaml: properly handle return value of yaml_pack_goto_state
The return value of this function was not handled. Fix this by
rethrowing it and adding its value to the "res" variable. To make this
a bit more readable, declare a macro to do that, in the same vein of
INDENT or PUT.

Change-Id: Icf33cb0befe39baa28bf586bfe19e301b924093a

rip-it: 89bf4e579056f70a2530f3a83e208c92a0024314