yaml: expose a yaml_pack_env_t object on pack API
Instead of having a single packing API, yaml_pack, force a
pack_env object to be created. This will allow adding more APIs to this
object, before pack is called:
* set_flags for specific packing flags
* set_output_directory for pack_file / subfiles
* etc

To keep the yaml_pack_env_t definition private, the object is now
t_allocated, which means that current_path as well as the error buffer
can be allocated on the t_scope as well.

Change-Id: I6df1b1dbcd04f8900e87de0d965f66044b85d5a9

rip-it: 11c26733cbaec20b4976d85c27b8b4c91ae56051