yaml: add API to set output directory of packing
This API can be called to specify the output directory where files will
be written. For the moment it is a bit useless, but it will be useful
when packing included subfiles, as well as to ensure we do not write
files outside of this output directory.

Change-Id: Iedb770d06a9c8c27e0d8f99a976ee9114ab6b8d8

rip-it: 49c4ab5dd5c47addaff8e827220c0b254e95e84d