yaml: fix packing of prefix comments with tagged scalar
If a scalar is tagged, the tag we put first, before the prefix comments.
This is backward to what we want, which is first the comments, then
the tag followed by the scalar.
The code is slightly worse now than before however :(

Change-Id: Ie34269dbb7664e58b4d1d3fb592f98afc57763f9

rip-it: 6d5476c5645bf34b95740deaaf8f61444ea7e1f6