yaml: fix double newlines at end of packed file
When we pack a file, we always add a newline at the end of the file.
However, when packing inline comments, we also always end with a
newline, to separate the comments from the rest. This caused a double
newline when a packed file ends with an inline comment.
This is easily fixed thanks to the packing state.

Change-Id: I544af39cbd6d39c0041cc13912317d924c370467

rip-it: f64a81fb73ec0c2f87df208152b44bb6ff74469f