yaml: reuse included presentation data when inlining contents
If we are not repacking into a file, included subfiles are not
recreated, but their contents inlined in the current stream. We want in
that case to still reuse the presentation data from the included
subfiles. However, this presentation data is set with paths relative to
the included data root, and not the including file's current path.

To properly reuse the presentation data, we need to set it in place of
the current presentation data, but also reset the current path.

Change-Id: I7cb18c69e2af610f6e845d24a9fda65f1fff3c84

rip-it: 69d71366ff45ef436ee3a3e941d89c49eb577af4